refactor(session): extract contract codec to _skill_session_contract_codec.py
Move validation, source/exploration-vector/execution-identity serialization,
contract to/from dict, manifest construction, and _digest_json helper from
_skill_session_contract_store.py (907 → 393 lines) into
_skill_session_contract_codec.py. The parent re-exports _digest_json and
all other moved symbols so existing imports keep working — notably
tests/server/test_run_skill_resume.py:640 and tests/execution/test_skill_session_contract_store.py
which import _digest_json from the canonical path.

Refs: #4664
